Rename fixture file
